Restructure project as Pebble lib
