+++ .gitignore
@@ -1,0 +1,9 @@
+# Ignore build generated files
+build/
+dist/
+
+# Ignore waf lock file
+.lock-waf*
+
+# Ignore installed node modules
+node_modules/
