+++ src/c/pge_splash.c
