+++ src/c/pge_sprite.c
